requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.31.1
beautifulsoup4==4.11.2
bottle==0.12.24
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.5
cycler==0.11.0
edge-tts==6.1.3
Eel==0.14.0
Flask==2.2.3
fonttools==4.37.4
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
gTTS==2.3.1
idna==3.4
importlib-metadata==6.0.0
imutils==0.5.4
itsdangerous==2.1.2
jaraco.context==4.3.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.0
more-itertools==9.1.0
mouse==0.7.1
MouseInfo==0.1.3
multidict==6.0.4
numpy==1.23.3
openai==0.27.0
opencv-python==4.6.0.66
packaging==21.3
pefile==2023.2.7
Pillow==9.2.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pygame==2.2.0
pygame-menu==4.4.3
PyGetWindow==0.0.9
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python-dotenv==1.0.0
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3
pywhatkit==5.4
pywin32==305
pywin32-ctypes==0.2.0
requests==2.28.2
six==1.16.0
soupsieve==2.4
SpeechRecognition==3.9.0
speedtest-cli==2.1.3
tqdm==4.64.1
typing_extensions==4.7.1
urllib3==1.26.14
Werkzeug==2.2.3
whichcraft==0.6.1
wikipedia==1.4.0
winspeech==1.0.1
wolframalpha==5.0.0
xmltodict==0.13.0
yarl==1.8.2
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
